DBMySQLManager: afegit classe Long als metòdes read i write. Classe PersonaDAO: checked all. Classe Persona: canviat persona_id de int a long (perquè a la BD és un INT UNSIGNED, que és més gran que un int de Java).

+++ src/model/DAO/DAODB.java
@@ -5,16 +5,25 @@ import model.Persona;
 import java.util.List;
 
 public interface DAODB <T, K> {
+
     // CRUD
     boolean create(T t);
+
     boolean read(T t);
+
     T readById (K id); // read by key (id)
+
     boolean update (T t);
+
     boolean update(T t, String ... camps);
+
     boolean delete(T t);
+
 
     // ALTRES
     boolean exists(T t);
-    int count();
+
+    long count();
+
     List<T> all();
 }
